server_py/requirements.txt
flask
flask-cors
python-dotenv
pymysql
bcrypt
pyjwt